all ids changed to int64

+++ internal/models/banks.go
@@ -2,7 +2,7 @@ package models
 
 type Bank struct {
 	// Internal id for system
-	Id int `db:"id" json:"id"`
+	Id int64 `db:"id" json:"id"`
 
 	// Readable name
 	Name string `db:"name" json:"name"`
