+++ internal/models/user_permissions.go
@@ -11,9 +11,9 @@ const (
 )
 
 type UserPermission struct {
-	Id                int         `db:"id"  json:"id"`
-	PermissionGroupId int         `db:"permission_group_id" json:"permission_group_id"`
-	UserId            int         `db:"user_id"  json:"user_id"`
+	Id                int64       `db:"id"  json:"id"`
+	PermissionGroupId int64       `db:"permission_group_id" json:"permission_group_id"`
+	UserId            int64       `db:"user_id"  json:"user_id"`
 	Level             AccessLevel `db:"level"   json:"level"`
 	CreatedAt         time.Time   `db:"created_at" json:"created_at"`
 }
